major file restructuring, put the files based on their scope instead of spread across
