+++ components/layout/layout.module.css
@@ -8,8 +8,4 @@
   height: 100%;
   padding: 4rem 0;
   flex: 1;
-  display: flex;
-  flex-direction: row;
-  justify-content: start;
-  align-items: center;
 }
