+++ components/my-certifications/my-certifications.module.css
@@ -1,0 +1,33 @@
+.certificationsWrapper {
+  padding: 4rem 0;
+}
+
+.certificationContainer {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  gap: 35px;
+}
+
+.certificationCard {
+  background: #3C3C3C;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  border-radius: 15px;
+  padding: 7%;
+}
+
+.certificationTitle {
+  font-weight: 700;
+  font-size: 18px;
+  line-height: 12px;
+  color: #FFFFFF;
+  margin-bottom: 1rem;
+}
+
+.certification {
+  margin-bottom: 0.75rem;
+}
+
+.certificationsWrapper {
+  padding: 4rem 0;
+}
