+++ components/my-skills/my-skills-card-types.ts
