+++ components/sidebar/sidebar.module.css
@@ -1,5 +1,6 @@
 .sidebar {
-  width: 20%;
+  width: 15%;
+  position: fixed;
 }
 
 .sidenavItem {
