+++ components/themed-tags/custom-h1/custom-h1.tsx
@@ -1,7 +1,6 @@
-import styles from '../styles/custom-h1.module.css';
+import styles from './custom-h1.module.css';
 
 type CustomH1Props= {
-  className: string;
   children: React.ReactNode;
 }
