+++ pages/contact.tsx
@@ -1,5 +1,5 @@
-import CustomH1 from '../components/custom-h1'
-import Layout from '../components/layout'
+import CustomH1 from '../components/custom-h1/custom-h1'
+import Layout from '../components/layout/layout'
 
 import styles from '../styles/Home.module.css'
 import contactStyles from '../styles/contact.module.css'
@@ -7,7 +7,7 @@ import contactStyles from '../styles/contact.module.css'
 export default function Contact() {
   return (
     <Layout>
-      <CustomH1 className={styles.title}>
+      <CustomH1>
           Email me :)
       </CustomH1>
       <div className={contactStyles.myEmail}>
