+++ pages/experience.tsx
@@ -1,12 +1,12 @@
-import CustomH1 from '../components/custom-h1'
-import Layout from '../components/layout'
+import CustomH1 from '../components/custom-h1/custom-h1'
+import Layout from '../components/layout/layout'
 
 import styles from '../styles/Home.module.css'
 
 export default function Experience() {
   return (
     <Layout>
-      <CustomH1 className={styles.title}>
+      <CustomH1>
           Hi,
           I am Ramu Narasinga
         </CustomH1>
