--- pages/my-skills.tsx
@@ -1,13 +1,0 @@
-import Layout from '../components/layout'
-import MyCertificationsContainer from '../components/my-certificatiions'
-import MySkillsContainer from '../components/my-skills-container'
-
-export default function MySkills() {
-
-  return (
-    <Layout>
-      <MySkillsContainer></MySkillsContainer>
-      <MyCertificationsContainer></MyCertificationsContainer>
-    </Layout> 
-  )
-}
