--- styles/my-certifications.module.css
@@ -1,3 +1,0 @@
-.certificationsWrapper {
-  padding: 4rem 0;
-}
